stock_analysis/requirements.txt
Django==1.11
django-cors-headers==2.4.0
djangorestframework==3.8.2
pytz==2018.5
xlrd==1.1.0
